[navigation 추가] 7. 네이밍 변경(activity -> fragment)
