+++ app/src/main/java/com/nexters/yetda/android/ui/detail/DetailFragment.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.Observer
 import androidx.navigation.fragment.navArgs
 import com.nexters.yetda.android.R
 import com.nexters.yetda.android.base.BaseFragment
-import com.nexters.yetda.android.databinding.ActivityDetailBinding
+import com.nexters.yetda.android.databinding.FragmentDetailBinding
 import com.nexters.yetda.android.domain.database.model.History
 import com.nexters.yetda.android.domain.database.model.Present
 import org.koin.android.viewmodel.ext.android.viewModel
@@ -15,10 +15,10 @@ import java.time.MonthDay
 import java.time.format.DateTimeFormatter
 import java.util.*
 
-class DetailActivity : BaseFragment<ActivityDetailBinding>() {
-    override val layoutResourceId = R.layout.activity_detail
+class DetailFragment : BaseFragment<FragmentDetailBinding>() {
+    override val layoutResourceId = R.layout.fragment_detail
     val viewModel: DetailViewModel by viewModel()
-    private val args: DetailActivityArgs by navArgs()
+    private val args: DetailFragmentArgs by navArgs()
 
     private val TAG = javaClass.simpleName
