+++ app/src/main/java/com/nexters/yetda/android/ui/gender/GenderFragment.kt
@@ -5,14 +5,14 @@ import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import com.nexters.yetda.android.R
 import com.nexters.yetda.android.base.BaseFragment
-import com.nexters.yetda.android.databinding.ActivityGenderBinding
+import com.nexters.yetda.android.databinding.FragmentGenderBinding
 import org.koin.android.viewmodel.ext.android.viewModel
 
 
-class GenderActivity : BaseFragment<ActivityGenderBinding>() {
-    override val layoutResourceId = R.layout.activity_gender
+class GenderFragment : BaseFragment<FragmentGenderBinding>() {
+    override val layoutResourceId = R.layout.fragment_gender
     val viewModel: GenderViewModel by viewModel()
-    val args: GenderActivityArgs by navArgs()
+    val args: GenderFragmentArgs by navArgs()
 
     private val TAG = javaClass.simpleName
 
@@ -26,7 +26,7 @@ class GenderActivity : BaseFragment<ActivityGenderBinding>() {
 
         viewModel.startNextActivityEvent.observe(this, Observer {
             findNavController().navigate(
-                GenderActivityDirections.actionGenderToBirthday(
+                GenderFragmentDirections.actionGenderToBirthday(
                     viewModel.name.value ?: "", viewModel.getGender()
                 )
             )
