+++ app/src/main/java/com/nexters/yetda/android/ui/member/MemberFragment.kt
@@ -3,10 +3,10 @@ package com.nexters.yetda.android.ui.member
 import android.widget.Toast
 import com.nexters.yetda.android.R
 import com.nexters.yetda.android.base.BaseFragment
-import com.nexters.yetda.android.databinding.ActivityMemberBinding
+import com.nexters.yetda.android.databinding.FragmentMemberBinding
 
-class MemberActivity : BaseFragment<ActivityMemberBinding>() {
-    override val layoutResourceId = R.layout.activity_member
+class MemberFragment : BaseFragment<FragmentMemberBinding>() {
+    override val layoutResourceId = R.layout.fragment_member
     private val TAG = javaClass.simpleName
 
     override fun initViewStart() {
