+++ app/src/main/java/com/nexters/yetda/android/ui/name/NameFragment.kt
@@ -4,12 +4,12 @@ import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
 import com.nexters.yetda.android.R
 import com.nexters.yetda.android.base.BaseFragment
-import com.nexters.yetda.android.databinding.ActivityNameBinding
+import com.nexters.yetda.android.databinding.FragmentNameBinding
 import org.koin.android.viewmodel.ext.android.viewModel
 
 
-class NameActivity : BaseFragment<ActivityNameBinding>() {
-    override val layoutResourceId = R.layout.activity_name
+class NameFragment : BaseFragment<FragmentNameBinding>() {
+    override val layoutResourceId = R.layout.fragment_name
     val viewModel: NameViewModel by viewModel()
 
     private val TAG = javaClass.simpleName
@@ -25,7 +25,11 @@ class NameActivity : BaseFragment<ActivityNameBinding>() {
         //        Crashlytics.getInstance().crash() // Force a crash
 
         viewModel.startNextActivityEvent.observe(this, Observer {
-            findNavController().navigate(NameActivityDirections.actionNameToGender(viewModel.name.value ?: ""))
+            findNavController().navigate(
+                NameFragmentDirections.actionNameToGender(
+                    viewModel.name.value ?: ""
+                )
+            )
         })
         viewModel.backBeforeActivityEvent.observe(this, Observer {
             findNavController().popBackStack()
