+++ app/src/main/res/layout/fragment_detail.xml
