+++ app/src/main/res/layout/fragment_question2.xml
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+
+        <variable
+            name="vm"
+            type="com.nexters.yetda.android.ui.question.QuestionViewModel" />
+    </data>
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="#03A9F4"
+        tools:context=".ui.question.QuestionFragment">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:gravity="center"
+            android:orientation="vertical">
+            <!--START cardView-->
+            <RelativeLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal">
+
+                <ImageView
+                    android:id="@+id/question_card_no"
+                    style="@style/iv"
+                    android:layout_centerVertical="true"
+                    android:padding="@dimen/click_padding"
+                    app:srcCompat="@android:drawable/ic_menu_close_clear_cancel" />
+
+                <ImageView
+                    android:id="@+id/question_card_view"
+                    style="@style/iv"
+                    android:layout_width="match_parent"
+                    android:layout_toStartOf="@+id/question_card_yes"
+                    android:layout_toEndOf="@+id/question_card_no"
+                    android:src="@drawable/ic_launcher_background" />
+
+                <ImageView
+                    android:id="@+id/question_card_yes"
+                    style="@style/iv"
+                    android:layout_alignParentEnd="true"
+                    android:layout_centerVertical="true"
+                    android:padding="@dimen/click_padding"
+                    app:srcCompat="@android:drawable/btn_radio" />
+
+            </RelativeLayout>
+
+            <TextView
+                android:id="@+id/text_idontknow"
+                style="@style/btn"
+                android:text="@string/idontknow" />
+
+        </LinearLayout>
+    </FrameLayout>
+
+</layout>
