+++ app/src/main/res/layout/fragment_splash.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.SplashActivity">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -16,6 +14,5 @@
             android:layout_height="wrap_content"
             app:lottie_autoPlay="true"
             app:lottie_rawRes="@raw/splash" />
-
     </LinearLayout>
 </layout>
