+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,8 +7,8 @@
 
     <fragment
         android:id="@+id/splash"
-        android:name="com.nexters.yetda.android.ui.SplashActivity"
-        tools:layout="@layout/activity_splash">
+        android:name="com.nexters.yetda.android.ui.SplashFragment"
+        tools:layout="@layout/fragment_splash">
         <action
             android:id="@+id/action_splash_to_home"
             app:destination="@id/home"
@@ -17,8 +17,8 @@
     </fragment>
     <fragment
         android:id="@+id/home"
-        android:name="com.nexters.yetda.android.ui.home.HomeActivity"
-        tools:layout="@layout/activity_home">
+        android:name="com.nexters.yetda.android.ui.home.HomeFragment"
+        tools:layout="@layout/fragment_home">
         <action
             android:id="@+id/action_home_to_name"
             app:destination="@id/name" />
@@ -31,16 +31,16 @@
     </fragment>
     <fragment
         android:id="@+id/name"
-        android:name="com.nexters.yetda.android.ui.name.NameActivity"
-        tools:layout="@layout/activity_name">
+        android:name="com.nexters.yetda.android.ui.name.NameFragment"
+        tools:layout="@layout/fragment_name">
         <action
             android:id="@+id/action_name_to_gender"
             app:destination="@id/gender" />
     </fragment>
     <fragment
         android:id="@+id/gender"
-        android:name="com.nexters.yetda.android.ui.gender.GenderActivity"
-        tools:layout="@layout/activity_gender">
+        android:name="com.nexters.yetda.android.ui.gender.GenderFragment"
+        tools:layout="@layout/fragment_gender">
         <action
             android:id="@+id/action_gender_to_birthday"
             app:destination="@id/birthday" />
@@ -51,8 +51,8 @@
     </fragment>
     <fragment
         android:id="@+id/birthday"
-        android:name="com.nexters.yetda.android.ui.birthday.BirthdayActivity"
-        tools:layout="@layout/activity_birthday">
+        android:name="com.nexters.yetda.android.ui.birthday.BirthdayFragment"
+        tools:layout="@layout/fragment_birthday">
         <action
             android:id="@+id/action_birthday_to_price"
             app:destination="@id/price" />
@@ -67,8 +67,8 @@
     </fragment>
     <fragment
         android:id="@+id/price"
-        android:name="com.nexters.yetda.android.ui.price.PriceActivity"
-        tools:layout="@layout/activity_price">
+        android:name="com.nexters.yetda.android.ui.price.PriceFragment"
+        tools:layout="@layout/fragment_price">
         <action
             android:id="@+id/action_price_to_question"
             app:destination="@id/question" />
@@ -87,8 +87,8 @@
     </fragment>
     <fragment
         android:id="@+id/question"
-        android:name="com.nexters.yetda.android.ui.question.QuestionActivity"
-        tools:layout="@layout/activity_question">
+        android:name="com.nexters.yetda.android.ui.question.QuestionFragment"
+        tools:layout="@layout/fragment_question">
         <action
             android:id="@+id/action_question_to_result"
             app:destination="@id/result" />
@@ -101,8 +101,8 @@
     </fragment>
     <fragment
         android:id="@+id/result"
-        android:name="com.nexters.yetda.android.ui.result.ResultActivity"
-        tools:layout="@layout/activity_result">
+        android:name="com.nexters.yetda.android.ui.result.ResultFragment"
+        tools:layout="@layout/fragment_result">
         <argument
             android:name="historyId"
             android:defaultValue="0"
@@ -116,15 +116,15 @@
 
     <fragment
         android:id="@+id/detail"
-        android:name="com.nexters.yetda.android.ui.detail.DetailActivity"
-        tools:layout="@layout/activity_detail" >
+        android:name="com.nexters.yetda.android.ui.detail.DetailFragment"
+        tools:layout="@layout/fragment_detail" >
         <argument
             android:name="history"
             app:argType="com.nexters.yetda.android.domain.database.model.History" />
     </fragment>
     <fragment
         android:id="@+id/member"
-        android:name="com.nexters.yetda.android.ui.member.MemberActivity"
-        tools:layout="@layout/activity_member" />
+        android:name="com.nexters.yetda.android.ui.member.MemberFragment"
+        tools:layout="@layout/fragment_member" />
 
 </navigation>
